phy, leds: add support for led triggers on phy link state change
This introduces an option CONFIG_LED_TRIGGER_PHY (default n), which will create
a set of led triggers for each instantiated PHY device.  There is one LED
trigger per link-speed, per-phy.

This allows for a user to configure their system to allow a set of LEDs to
represent link state changes on the phy.

Signed-off-by: Josh Cartwright <josh.cartwright@ni.com>