zynqproto: adjust device tree to reflect actual TTC clock-frequency.
ARM_PLL is configured to generate 1333.33MHz from the 33.3333MHz osc.
ARM_CLK_CTRL is programmed to divide ARM_PLL by 2, with CLK_621_TRUE enabled.

This leaves the frequency table:
      CPU_6X4X:   666.667Mhz
      CPU_3X2X:   333.333MHz
      CPU_2X:     222.222MHz
      CPU_1X:     111.111MHz

The TTC is driven by CPU_1X, which means the device tree should
reflect 111.111MHz as its clock-frequency.

Since the first of the TTC counters is used as the clocksource and sys_timer,
this also fixes observed clock drift.

Signed-off-by: Josh Cartwright <josh.cartwright@ni.com>