xuartps: prevent disabling of tx during xuartps_shutdown
When xuartps_shutdown is called, it disables both the RX/TX portions of the
uart.  This is problematic in the case where usermode has no handle open to
ttyPS0, and the write callback of the console driver is told to flush the
printk buffer.

This case was hit in some timing scenarios on bootup w/ PREEMPT-RT.  The
problematic scenario looked something like this pseudobacktrace:

  open("/dev/ttyPS0")
    ... can be nested ...
  close("/dev/ttyPS0")
    fput()
      tty_release()
         xuartps_shutdown()
           ... TX/RX both disabled ...
  open("/dev/tty")
    console_lock()
    console_unlock()
      _call_console_drivers()
        xuartps_console_write()
          xuartps_console_wait_tx()
             ... spin indefinitely, waiting for TXEMPTY (but condition never
                 held because there are bytes in the FIFO and TX is disabled) ...

With this change, the TX logic is never disabled when usermode closes a handle
to ttyPS0.  This lets the TX FIFO empty after shutdown, and leaves the uart in
a state where the console can use it, even if no user handle is open.

Signed-off-by: Josh Cartwright <josh.cartwright@gmail.com>