xilinx_uartps: Use the channel status register when checking for TX empty.
When draining the TX FIFO, the xilinx_uartps driver currently uses the interrupt
status register to determine when the FIFO is empty. The ISR, though, only
indicates that an event has happened since we last checked. The interrupt
service handler can clear and mask off TX empty events from the ISR. The channel
status register is what indicates the actual current state. This patch tells the
driver to use the channel status register instead of the ISR when checking for
TX empty.

Signed-off-by: Corbin Atkinson <corbin.atkinson@ni.com>
Acked-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Jason Smith <jason.smith@ni.com>