Zynq NAND: Fixed some problems with our timing values
The code that reads the timing values from the device tree wasn't converting
the values from big-endian, so we were writing an invalid value to the SMC
timing register. The SMC ignored this invalid value, and the SMC continued
working just fine, but we were unknowingly using the timing values programmed
in the FSBL. Updated this code to convert the values from big-endian.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>