Marvell 88E151x PHY: Configure interrupt output
By default, the interrupt output of the Marvell 88E1512 PHY is
configured as an LED output, and is driven low. If we have a case where
eth1 is brought up before eth0, this causes an interrupt storm, and
Linux eventually gives up and disables the PHY interrupt line. Adding a
probe function to the Marvell 88E1512 PHY driver and moving the
interrupt output configuration there resolves this problem.

[bm: Adjust for d2fa47d9dd5c5f82d78c4503cf33989f67e8aa64 'phy: marvell: Add ethtool statistics counters']

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>
Signed-off-by: Mihaly Varga <mihaly.varga@ni.com>
Acked-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Jaeden Amero <jaeden.amero@ni.com>
Natinst-ReviewBoard-ID: 81627

[bm: Restore correct page on LED/irq config]
Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Jeff Westfahl <jeff.westfahl@ni.com>
Natinst-ReviewBoard-ID: 148255

[gratian: split LED pin setup from m88e151x_probe; fixes conflict with 0b04680fdae4 ("phy: marvell: Add support for temperature sensor")]
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>