Marvell 88E151x PHY: Configure interrupt output
By default, the interrupt output of the Marvell 88E1512 PHY is
configured as an LED output, and is driven low. If we have a case where
eth1 is brought up before eth0, this causes an interrupt storm, and
Linux eventually gives up and disables the PHY interrupt line. Adding a
probe function to the Marvell 88E1512 PHY driver and moving the
interrupt output configuration there resolves this problem.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>
Signed-off-by: Mihaly Varga <mihaly.varga@ni.com>
Acked-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Jaeden Amero <jaeden.amero@ni.com>
---
Natinst-ReviewBoard-ID: 81627