Zynq USB: Fix modularization problems and modularize NI Zynq proto defconfig
Changed Kconfig of USB_GADGET_XUSBPS from boolean to tristate so that the
Zynq USB peripheral driver can be built as a module.

Added a description to the Kconfig of USB_XUSBPS_DR_OF so that we can
choose to build it into the kernel when the rest of the Zynq USB drivers
are built as modules. The Zynq USB drivers don't work as modules unless
this piece is built into the kernel. Fixed USB Makefile for this as well.

Export GPL symbol usb_port_suspend from usb/core/hub.c. The Zynq OTG driver
uses this function.

Updated NI Zynq proto defconfig to completely modularize USB.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>