atmel_serial: Use atmel_port consistently
In all functions other than atmel_serial_probe, the name "port" is
used to refer to an instance of struct uart_port. In many of these
functions, "atmel_port" is used to refer to an instance of struct
atmel_uart_port.

We make the use of the name "port" consistent by making
atmel_serial_probe use "atmel_port" to refer to an instance of struct
atmel_uart_port instead of its previous name of "port".

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>