Zynq Ethernet: Fix problem determining correct transmit clock register
There's some kind of offset for interrupt numbers in the device tree with
our current trunk. There was code in there reading the interrupt number
directly from the device tree instead of going through OF, so it got the
wrong number and chose the wrong register. We've already read the interrupt
from OF at that point, so fixed it to just use the correct value that we've
already read.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>