irq: Honor /proc/irq core affinity when IRQ thread is created
Currently, by writing to /proc/irq/../smp_affinity file, core affinity of
already running IRQ threads can be set. However, after writing to the
/proc file, an IRQ thread that gets created does not inherit the affinity
specified in the file. This happens because the irq action is registered
in irq description only after setup_affinity function is called. This patch
addresses this problem by calling setup_affinity function after the irq
action is updated in the irq description structure. It looks like the existing
behavior (without the patch) is also inconsistent because in the case of
multiple action handlers for an irq, the IRQ thread associated with the new
action handler that is being registered does not inherit the core affinity
from the /proc file, however, the IRQ threads of the existing action handlers
do inherit the affinity.