Zynq Ethernet: Better Tx watchdog implementation
Implemented a different Tx watchdog that expires when the earliest
unprocessed Tx packet has gone beyond the timeout value. With our
long Tx ring, it can take a long time for the default watchdog to
expire because Tx packets can be queued successfully until we run
out of Tx buffers.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>