Zynq Ethernet: Make Ethernet through the PL work on DosX
Updated DosX device tree to use FPGA_CLK0. FPGA_CLK1 appears to be used
for the UARTs.

Added support for an EMIO GPIO for telling the GMII-RGMII conversion logic
in the PL what the link speed is. This is necessary for 1Gb Ethernet to
work.

Also removed LSI ET1011C PHY support on DosX. Only the prototype has this
PHY.

These changes also magically fix the secondary Ethernet speed LED.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>