Zynq Ethernet: Synchronization fixes
Completely removed the spinlock from the Zynq Ethernet driver. Based on an
examination of other Linux Ethernet drivers, for example the Intel PRO/1000
driver at drivers/net/ethernet/intel/e1000, this lock isn't necessary. The
network operations generally are synchronized externally by higher levels
of the network stack. For example, ndo_open and ndo_stop are called with
the RTNL lock held, and ndo_start_xmit is called with a transmit queue
lock held. Notice that the receive path in the Zynq driver has not ever
been synchronized with the spinlock, and it appears to function just fine.

Updated the transmit queue free count to be atomic_t for synchronization
between xemacps_start_xmit and xemacps_tx_poll. Updated queueing of the
transmit task to use the system non-reentrant queue. I mistakenly thought
that work_pending would return false until the task finished running but
this is not the case. Removing the spinlock allowed the the transmit task
to run more than once at the same time, causing intermittent crashes. Using
the non-reentrant work queue solves this problem.

Added some state management to keep track of when the interface is down and
when it's being reset. Using netif_carrier_ok for this is not appropriate
because that's for link state.

Set netif_carrier_on only when we've detected link up from the PHY.

Don't set netif_carrier_off when we're just resetting the interface. This
speeds up the reset recovery.

Use rtnl_lock/rtnl_unlock to synchronize access to the polling mode sysfs
variables.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>