i2c, xilinx: Fixed order of operations in xi2cps_mrecv()
Moved writing the I2C slave address after the calls for setting
the transfer size and enabling interrupts.
Writing the slave address initiates the I2C transfer and we can
get into a situation where the transfer size register contains an
invalid value and the chip clocks in more bytes than requested
resulting in a FIFO overflow and hang.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>