Zynq Ethernet: Fix for rare crash during reset
In xemacps_down, we need to make sure xemacps_tx_task is no longer running
or scheduled before we can successfully stop calls to xemacps_xmit_start
with netif_stop_queue, because xemacps_tx_task will restart the queue if
it's been stopped. We've seen a rare crash where xemacps_xmit_start is
called in the middle of the reset process.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>