irq: reduce latency for first interrupt in irq thread in some cases
By setting the affinity in the irq thread before it waits for interrupts
to occur, the additional latency that occurs to set the affinity can
be avoided for handling the first interupt that occurs for an irq. This
could especially be useful to applications that do not change the
affinity dynamically.