irq: Add priority support to /proc/irq/../
This patch allows configuring priority for different irq threads through
the /proc/irq/ system (much same as the existing mechanism to configure
the core affinity for irqs). Priority of an already running irq thread
can be changed also, however, the new priority will take into effect
when the irq thread is scheduled to run the next time. An irq thread
that gets created will use the priority specified in the
/proc/irq/../priority file. Thanks Josh Cartwright for suggesting this
idea.

Signed-off-by: Sankara S Muthukrishnan <sankara.m@ni.com>
Acked-by: Josh Cartwright <josh.cartwright@ni.com>