ni16550: Don't implement RS-485 IOCTLs for RS-232
User mode needs a way to tell the difference between an RS-232 and an
RS-485 port. The traditional way to do this is to try to call an
RS-485 specific IOCTL on the port and have it succeed only if the port
is RS-485.

Since the UART driver doesn't know how hardware is hooked up to its
UART, we use device tree to tell of_serial whether or not to register
with the NI 16550 RS-485 transceiver driver.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>