!!! fix priority check issues
This is to fix commit 412c3c51320a120d612469836339963e1ede3a1a, in which
the main issue was the modification in the function rt_prio in
include/linux/sched.h. rt_prio() is called with normal priority (which
is MAX_RT_PRIO-1 - rt_priority) internally in scheduler functions. In
other words, rt_prio() should not be called with rt_priority (which was
done in the problematic commit).  This patch reverts that change and
leaves rt_prio() same as it was and thus fixes the issue reported in CAR
376725.

Then, this patch validates the rt_priority given as input by user (
through kernel boot parameter or /proc/irq/../priority mechanism)
correctly without using rt_prio().

Signed-off-by: Sankara S Muthukrishnan <sankara.m@ni.com>
Acked-by: Joe Hershberger <joe.hershberger@ni.com>
Acked-by: Scot Salmon <scot.salmon@ni.com>
Acked-by: Josh Cartwright <josh.cartwright@ni.com>
Acked-by: Jeff Westfahl <jeff.westfahl@ni.com>