Revert "irq: Honor /proc/irq core affinity when IRQ thread is created"
This reverts commit 0d63820107f809ba7a84b1413e5d1c442e638707.  This also
reverts commit f2ec6beaa139028048caead9a9e0bf0c6ad1adc2 partially. The
changes are reverted because the mechanism to handle irq affinity and
priority does not work for shared interrupts. Commit
04aa530ec04f61875b99c12721162e2964e3318c already applied in mainline by
Thomas Gleixner will be applied on top of this (which works for both
shared and non-shared interrupts).

Signed-off-by: Sankara S Muthukrishnan <sankara.m@ni.com>
Acked-by: Josh Cartwright <josh.cartwright@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>