Zynq Ethernet: Prevent crash when removing a previously enabled device
If you attempt to remove a Zynq Ethernet device that has been enabled (for
example by "ifconfig eth0 up"), and you then try to remove that device from
the system (for example by loading our EtherCAT driver), the Zynq Ethernet
driver crashes. We can easily fix that by NULLing out the PHY device when
the interface is closed. Note that we don't see that with our more normal
EtherCAT use case because the secondary port is never enabled for Ethernet
in that case.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>