net: fix misaligned access in ipv4 and ipv6
This patch workarounds misaligned memory accesses that happen in kernel
mode from ipv4 and ipv6 stack. The test cases to reproduce the
misaligned kernel access were to do IPv4 and IPv6 pings to an ARM Linux
target(zynq). On a Cortex A9 processor, instructions such as load
multiple and store multiple that do misaligned memory access cause a
fault that affects performance (fault handling causes a latecny of 700
to 800 ns).

Signed-off-by: Sankara S Muthukrishnan <sankara.m@ni.com>
Acked-by: Josh Cartwright <josh.cartwright@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>