net: xilinx emacps: Share a page for transmit and receive descriptors
Coherent DMA allocations are done in pages. Since we get a whole page, we
should use it, and we shouldn't waste coherent memory allocations when we
can combine transmit and receive descriptors into a single page and still
have more than enough. This change allocates 512 8-byte descriptors and
allocates 64 for transmit and 448 for receive. The total memory usage of
the driver goes down a bit even though we now have a much longer receive
ring.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>