net: xilinx emacps: Transmit rewrite.
Rewrote the whole transmit side of the driver. We no longer use interrupts
for transmit descriptor clean up. Cleaning is now done actively in
xemacps_start_xmit. If the transmit descriptor ring fills up, we schedule
a job to clean things up. If we don't clean any descriptors after a second
we reset the interface. There's a timer that will run and clean transmit
descriptors if no more packets are transmitted, so that we don't have any
completed SKBs sitting around unfreed.

The transmit stall HW bug is worked around in a completely different way.
If we're running on a rev of Zynq that has this bug, we throttle transmits
such that only one packet can be sent at a time. The condition that causes
the bug requires multiple descriptors to be in use at once.

With this rewrite, I had three Zynqs run iperf over a weekend. Two of the
Zynqs are older chips with the transmit stall bug, and one is a rev 3.0
Zynq without the transmit stall bug. Combined (rx + tx) throughput of these
Zynqs was 378.5Mb/s, 381Mb/s, and 401Mb/s. Without with rewrite, I've never
seen iperf run successfully even overnight.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>