net: xilinx emacps: Fix receive packet loop
The last commit failed to update the next receive descriptor addr value
at the end of the loop. This caused the loop to terminate after processing
only one receive packet. This wasn't a super big deal because the outer
loop would keep calling until the NAPI budget was filled. This change
just adds the necessary update to addr.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>