net: xemacps: Check for receive packets with bad FCS and ignore them
We've seen a very rare occurrence of receive frames that come in with an
apparently invalid receive descriptor control word. The control word looks
something like 0x0040A800, where the length field varies but the rest of
the value remains the same. This causes a BUG in skb_put because the length
field is larger than our receive buffer size. The bit 0x00002000 which is
set in this case should not ever be set unless FCSIGNORE is set in the
net_cfg register. This bit is not set, yet we get a packet with the bad
FCS bit set in the control word.

I did some testing and the receive DMA is not doing something terrible
like writing past the end of the receive buffer. That's good. Since this
seems to be otherwise benign, we can just check for the bad FCS bit in
the control word and ignore the packet.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>
Signed-off-by: Xander Huff <xander.huff@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Mihaly Varga <mihaly.varga@ni.com>
---
Natinst-ReviewBoard-ID: 82089