net: xilinx emacps: Added receive stall handler
Added a timer that checks if we've seen a receive error and haven't
received any good packets for a while. If this happens, we may have
triggered the bug. If we see this condition, we toggle RXEN, which
starts up packet reception again. A receive error of some kind is a
precondition to triggering the receive stall hardware bug.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>