net: xilinx emacps: Optimize switching between interrupt and polling modes
When switching between interrupt and polling modes, we were just scheduling
a full interface reset. However, this interface reset has side effects and
we should not use it if we don't have to. I updated the switching mechanism
to just affect the interrupt handler and polling thread. Switching between
modes works exactly as before, but without the side effects of a full
interface reset.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>