8250: Prevent deadlock during FPGA programming
When the FPGA goes down for reprogramming, the
serial driver should suspend the port and then
lock on fpga_lock. However if this order is reversed,
it could cause a deadlock since the suspend code
path also grabs fpga_lock. A similar situation will
arise when the FPGA comes back up after
reprogramming.

This commit fixes the order of operations which
takes place when the FPGA down/up notifications come
to the serial driver.

Signed-off-by: Karthik Manamcheri <karthik.manamcheri@ni.com>
Acked-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Kyle Teske <kyle.teske@ni.com>
Acked-by: Jason Smith <jason.smith@ni.com>