8250: Fix potential hang after FPGA reprogram fail
Currently, the functions which need access to
the UART hardware have to grab a fpga_lock and
check for the fpga_state. In the case when the
reprogramming fails, we do not unlock fpga_lock.
This would cause a subsequest retry on FPGA
reprogramming to hang.

This is fixed by making sure we release the
fpga_lock even when the reprogramming fails.

Signed-off-by: Karthik Manamcheri <karthik.manamcheri@ni.com>