usb: xilinxps_udc: Added support for Test Mode feature
The test mode is used for compliance testing of USB devices.
All the devices must support this feature for compliance.

As part of this feature, the host sends SET_FEATURE command
with TEST_MODE feature and device needs to enable that test mode.
The exit from TEST_MODE is to power cycle the device.

There are 5 test modes as given below.
1. Test_SE0_NAK
2. Test_J
3. Test_K
4. Test_Packet
5. Test_Force_Enable

For more information, refer to "section 7.1.20 Test Mode Support" of USB 2.0 spec.

Signed-off-by: Naveen Mamindlapalli <naveenm@xilinx.com>