8250: Do not try to grab fpga_lock too early
In uart_startup, we try to grab the fpga_lock
at the start of the function even before we check
if the port is valid and known. This could cause
problems if the port has an error. We should move
the grabbing of fpga_lock to after all the error
checking on the port has been done.

Signed-off-by: Karthik Manamcheri <karthik.manamcheri@ni.com>
Acked-by: Josh Cartwright <josh.cartwright@ni.com>
Acked-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Jason Smith <jason.smith@ni.com>