wl12xx: Make wl12xx driver depend on WIRELESS_EXT so the existing nitargetcfg software can use it.
This is a hack to get CONFIG_WIRELESS_EXT enabled, since if
it is turned on in the defconfig it will not propogate to
the .config (as it is not user selectable).  Even though the
driver does not need it, the current NI user-mode software
stacks require this support for scanning and interface
enumeration.  With this change, it will propogate to the
wl12xx driver build (from compat-wireless) and thus include
WEXT support in that driver and wireless networking modules.

Link: http://review-board.natinst.com/r/37925/
Signed-off-by: Sundar Subbiah <sundar.subbiah@ni.com>
Acked-by: James Minor <james.minor@ni.com>
Acked-by: Josh Cartwright <josh.cartwright@ni.com>
Acked-by: Jeff Westfahl <jeff.westfahl@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>