wlcore: Depend on WIRELESS_EXT for nitargetcfg
This is a hack to get CONFIG_WIRELESS_EXT enabled, since if it is turned
on in the defconfig it will not propogate to the .config (as it is not
user selectable).  Even though the driver does not need it, the current
NI user-mode software stacks require this support for scanning and
interface enumeration.  With this change, it will propogate to the
wl12xx driver build (from compat-wireless) and thus include WEXT support
in that driver and wireless networking modules.

Signed-off-by: Sundar Subbiah <sundar.subbiah@ni.com>
Acked-by: James Minor <james.minor@ni.com>
Acked-by: Josh Cartwright <josh.cartwright@ni.com>
Acked-by: Jeff Westfahl <jeff.westfahl@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>
---
Natinst-ReviewBoard-ID: 37925