8250: Do not try to remove an unregistered port
In serial8250_register_port, we should remove
a port only if the port is already registered with
the tty layer. Otherwise, we'll get a warning
saying that we are trying to remove a wrong port.

This is already fixed in the commit "8250_pnp: do
pnp probe before legacy probe". I have back-ported
the required change from that commit so that we
can suppress the warning message we see. Once the
newer commit has been merged, this commit can be
removed. Refer tree linux/kernel/git/torvalds/linux.git,
commit 835d844d1a28efba81d5aca7385e24c29d3a6db2.

Signed-off-by: Karthik Manamcheri <karthik.manamcheri@ni.com>
Acked-by: Jaeden Amero <jaeden.amero@ni.com>