drm/i915: rip out indirection for pfit pipe_config assignment
This was still required a bit (on the cargo-cult side though) when the
state was stored in dev_priv, and when the enable/disable sequence was
botched a bit (to avoid too many updates).

But with pipeconfig we always get a clean slate, so this is pointless.
Rip it out.

Reviewed-by: Mika Kuoppala <mika.kuoppala@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit 2deefda541edb0c73e57e988ccaac4cd014da0d3)
Signed-off-by: Darren Hart <dvhart@linux.intel.com>