i2c: stu300: do not request a specific clock name
We have used the default clock associated with the block
for a long time, only heuristics in the clock system has
made this work anyway. This needs to be done away with as
we start probing this driver and its clocks exclusively
from the device tree.

Acked-by: Wolfram Sang <wsa@the-dreams.de>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
(cherry picked from commit 2165f836c8f7036491fae41e9bc327a3cdf2fea3)
Signed-off-by: Darren Hart <dvhart@linux.intel.com>