wlcore/wl12xx: check if we got correct clock data from DT
The fref and the tcxo clocks settings are optional in some platforms.
WiLink8 doesn't need either, so we don't check the values.  WiLink 6
only needs the fref clock, so we check that it is valid or return with
an error.  WiLink7 needs both clocks, if either is not available we
return with an error.

Signed-off-by: Luciano Coelho <coelho@ti.com>
Acked-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: James Minor <james.minor@ni.com>
---
Natinst-ReviewBoard-ID: 84972