ARM: shmobile: Add koelsch defconfig
This is intended to be used until multi-arch is able to be
used for the koelsch board at which time a more generic configuration
will be used in place of this one.

This is based on the lager defconfig.

Signed-off-by: Simon Horman <horms+renesas@verge.net.au>
(cherry picked from commit 6bb2e347899b297eea26f2658eac8fdcd633fba5)
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>