nati_x86_64_defconfig: clean up and update for Intel Baytrail
Our previous nati_x86_64_defconfig contained a lot of extra/unused features.
This causes long build times and resulted in a kernel that was 5.4MB disk size,
7.3MB RAM footprint and 145.3MB of modules. It was also missing some of the
Intel Baytrail specific configurations.

Reconfigured the kernel to include the Intel Baytrail drivers and to contain
a minimal set of features similar to our Zynq defconfig. The idea behind this
was to start with a minimal set of features and decide later what niceties we
want to include. The resulting kernel has 3.9MB disk size, 4.9MB memory size
and 6.3MB total modules (plus improved build time).

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Brad Mouring <brad.mouring@ni.com>
---
Natinst-ReviewBoard-ID: http://review-board.natinst.com/r/54288