ni: x86_64: Add NI Watchdog driver
Added an NI Watchdog driver. This is an ACPI device that exposes the NI
Watchdog hardware interface.

Not all of the proposed features of the device work as expected.
Development work on this device by the hardware team is currently
not a high priority. These issues will be addressed once the hardware
team gets back to this device.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>