defconfig: Build the Intel 82575/82576 Gigabit Ethernet driver as a module
Change the Intel(R) 82575/82576 PCI-Express Gigabit Ethernet driver from an in
kernel driver to a module. We want this driver to be compiled as a module so
that it could be replaced by a specialized driver for Industrial Communications.

As an unintended (but desired) side effect PTP support is enabled. This fixes
another requirement to enable PTP support in our configs.

Natinst-CAR-ID: 442407
Natinst-CAR-ID: 433419

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>
Acked-by: Bill Pittman <bill.pittman@ni.com>
Acked-by: Brad Mouring <brad.mouring@ni.com>