nati_zynq_defconfig: Remove event debugging from kernel
This is supposed to be a module at best and on pretty much any
distribution that includes it it is blacklisted.  It causes any input
event to be logged to dmesg.  On any machine with a keyboard it is a
huge security hole because it fills the logs with every key stroke
(including passwords).  We have no need for this.  It was reported by a
customer to be filling the disk.

https://usfirst.collab.net/sf/go/artf2387

Signed-off-by: Joe Hershberger <joe.hershberger@ni.com>