sdhci: Add quirk and device tree parameter to force SD test mode
On some devices, CD is broken so that we must force the SDHCI into test
mode and set CD, so that it always detects an SD card as present.

In order to get myRIO TiWi working, we had previously always set the
SDHCI into test mode. Unfortunately, this had the side effect of making
all SD cards used with our NI Linux kernels undetectable and
non-removable.

By making this "SD test mode" setting optional via a quirk, we can avoid
this side effect for devices other than the myRIO.  Additionally, we add
a device parameter to sdhci-pltfm to allow all SDHCI drivers to enable
this quirk.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>

Conflicts:
	include/linux/mmc/sdhci.h