igb: support device_poll and netdev_poll interfaces
Add support for the new device_poll and netdev_poll interfaces to the igb
driver. This allows igb devices to be used without generating any hardware
interrupts, with minimal changes to the driver.

If the device_poll and netdev_poll features are included in mainline
or RT, this change to the igb driver could be included as a reference
implementation.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>