nizynqcpld: Use correct register to detect switches
The safemode, ipreset, and consoleout switches were being read from the
DebugSwitchReg, which only represents the jumpers and not the reset
button.  Also, on targets like myRIO and Sol, the bitfields in that
register changed from 0xE0 to 0x07.

The bitfields in the ProcessorMode register have never changed and
represent the actual state used by the bootloader.  Use it here too.

Signed-off-by: Joe Hershberger <joe.hershberger@ni.com>